rag-with-recipes/weaviaterecipes/pyproject.toml
[tool.poetry]
name = "weaviaterecipes"
version = "0.1.0"
description = ""
authors = ["Connor Shorten <connor@weaviate.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
dspy-ai = "^2.4.9"
cohere = "^5.5.4"
weaviate-client = "^4.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
